refactor: lift proxyRpcRequest into PackageViewServer base
