+++ src/providers/scout/scout-view-server.ts
@@ -6,32 +6,6 @@ import { AbsolutePath, toAbsolutePath } from "../../core/path";
 import { basename, dirname } from "../../core/uri";
 import { scoutBinPath } from "../../scout/props";
 
-// Custom request/response types for JSON-RPC proxy communication.
-// We can't use fetch's Request/Response/Headers because:
-// - They're not serializable (contain methods, streams, etc.)
-// - Headers is a class, not Record<string, string>
-// - Response.body is ReadableStream, not string
-// - We need bodyEncoding to indicate base64 for binary data
-//
-// Limitations vs native fetch:
-// - No streaming: bodies must be fully buffered as strings
-// - Binary data requires base64 encoding (adds ~33% overhead)
-// - Multi-value headers (e.g. Set-Cookie) collapse to single string
-// - Large request bodies must fit in memory
-export interface HttpProxyRpcRequest {
-  method: "GET" | "POST" | "PUT" | "DELETE";
-  path: string;
-  headers?: Record<string, string>;
-  body?: string;
-}
-
-export interface HttpProxyRpcResponse {
-  status: number;
-  headers: Record<string, string>;
-  body: string | null;
-  bodyEncoding: "utf8" | "base64";
-}
-
 export class ScoutViewServer extends PackageViewServer {
   public readonly legacy: {
     getScans: (results_dir?: Uri) => Promise<string>;
@@ -154,41 +128,5 @@ export class ScoutViewServer extends PackageViewServer {
     return (
       await this.api_json(`/api/v2/scans/${base64Dir}/${scanFile}`, "DELETE")
     ).data;
-  }
-
-  /**
-   * JSON-RPC method handler that proxies webview HTTP requests to the backend.
-   * Converts HttpProxyRpcRequest to fetch call, then serializes response for JSON-RPC transport.
-   */
-  async proxyRpcRequest(
-    request: HttpProxyRpcRequest
-  ): Promise<HttpProxyRpcResponse> {
-    await this.ensureRunning();
-
-    const { status, headers, data } = await this.serverFetch(
-      request.path,
-      request.method,
-      new Headers(request.headers),
-      request.body
-    );
-
-    const responseHeaders: Record<string, string> = {};
-    headers.forEach((value, key) => {
-      responseHeaders[key] = value;
-    });
-
-    return {
-      status,
-      headers: responseHeaders,
-      ...(data instanceof Uint8Array
-        ? {
-            body: Buffer.from(data).toString("base64"),
-            bodyEncoding: "base64",
-          }
-        : {
-            body: data,
-            bodyEncoding: "utf8",
-          }),
-    };
   }
 }
